refactor: delete comments, console, move some methods to view
