--- async-race/src/components/controller/controllerWinners.ts
@@ -1,11 +1,0 @@
-// import Api from '../model/api';
-
-// class ControllerWinners {
-//   private api: Api;
-
-//   constructor() {
-//     this.api = new Api();
-//   }
-// }
-
-// export default ControllerWinners;
